intent-nexus/django_backend/accounts/urls.py

from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView
from . import views

urlpatterns = [
    path('register/', views.RegisterView.as_view(), name='register'),
    path('login/', views.login_view, name='login'),
    path('profile/', views.profile_view, name='profile'),
    path('preferences/', views.UserPreferencesView.as_view(), name='preferences'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]
